Split the reference-filtering code into a separate file
